subresource-integrity/subresource-integrity.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Subresource Integrity</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<div id="container"></div>
<script>
  if ('serviceWorker' in navigator) {
    var workerName = 'service-worker.js';
    var serviceWorkerResult;
    register();
    function register() {
      var test = async_test('Service worker loading');
      return navigator.serviceWorker.register(workerName, {scope: './'}).then(function(registration) {
        return registration.unregister();
      }).then(function () {
        return navigator.serviceWorker.register(workerName, {scope: './'});
      }).then(function () {
        test.step(function() {
          assert_true(true, "Service worker loaded");
        });
        return test.done();
      }).catch(function(error) {
        test.step(function() {
          assert_unreached("Service worker not loaded")
        });
        return test.done();
      });
    }
  }
</script>

<script>
    // <script> tests
    var xorigin_anon_script = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-anon-script.js';

    var xorigin_creds_script = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-creds-script.js';

    var xorigin_ineligible_script = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-ineligible-script.js';

    var SRIScriptTest = function(pass, name, src, integrityValue, crossoriginValue, successLoadCallback) {
        this.pass = pass;
        this.name = "Script: " + name;
        this.src = src;
        this.integrityValue = integrityValue;
        this.crossoriginValue = crossoriginValue;
        this.successLoadCallback = successLoadCallback || function () {};
    }

    SRIScriptTest.prototype.execute = function() {
        var that = this;
        var test = async_test(this.name);
        var e = document.createElement("script");
        e.src = this.src;
        e.setAttribute("integrity", this.integrityValue);
        if(this.crossoriginValue) {
            e.setAttribute("crossorigin", this.crossoriginValue);
        }
        if(this.pass) {
            e.addEventListener("load", function() {
              that.successLoadCallback();
              test.done()
            });
            e.addEventListener("error", function() {
                test.step(function(){ assert_unreached("Good load fired error handler.") })
            });
        } else {
           e.addEventListener("load", function() {
                test.step(function() { assert_unreached("Bad load succeeded.") })
            });
           e.addEventListener("error", function() {test.done()});
        }
        document.body.appendChild(e);
    };

    var xorigin_anon_style = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-anon-style.css';

    var xorigin_creds_style = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-creds-style.css';

    var xorigin_ineligible_style = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-ineligible-style.css';

    // <link> tests
    // Style tests must be done synchronously because they rely on the presence
    // and absence of global style, which can affect later tests. Thus, instead
    // of executing them one at a time, the style tests are implemented as a
    // queue that builds up a list of tests, and then executes them one at a
    // time.
    var SRIStyleTest = function(queue, pass, name, attrs, customCallback, altPassValue) {
        this.pass = pass;
        this.name = "Style: " + name;
        this.customCallback = customCallback || function () {};
        this.attrs = attrs || {};
        this.passValue = altPassValue || "rgb(255, 255, 0)";

        this.test = async_test(this.name);

        this.queue = queue;
        this.queue.push(this);
    }

    SRIStyleTest.prototype.execute = function() {
        var that = this;
        var container = document.getElementById("container");
        while (container.hasChildNodes()) {
          container.removeChild(container.firstChild);
        }

        var test = this.test;

        var div = document.createElement("div");
        div.className = "testdiv";
        var e = document.createElement("link");
        this.attrs.rel = this.attrs.rel || "stylesheet";
        for (var key in this.attrs) {
            if (this.attrs.hasOwnProperty(key)) {
                e.setAttribute(key, this.attrs[key]);
            }
        }

        if(this.pass) {
            e.addEventListener("load", function() {
                test.step(function() {
                    var background = window.getComputedStyle(div, null).getPropertyValue("background-color");
                    assert_equals(background, that.passValue);
                    test.done();
                });
            });
            e.addEventListener("error", function() {
                test.step(function(){ assert_unreached("Good load fired error handler.") })
            });
        } else {
            e.addEventListener("load", function() {
                 test.step(function() { assert_unreached("Bad load succeeded.") })
             });
            e.addEventListener("error", function() {
                test.step(function() {
                    var background = window.getComputedStyle(div, null).getPropertyValue("background-color");
                    assert_equals(background, "rgba(0, 0, 0, 0)");
                    test.done();
                });
            });
        }
        container.appendChild(div);
        container.appendChild(e);
        this.customCallback(e, container);
    };

    var style_tests = [];
    style_tests.execute = function() {
        if (this.length > 0) {
            this.shift().execute();
        }
    }
    add_result_callback(function(res) {
        if (res.name.startsWith("Style: ")) {
          style_tests.execute();
        }
    });
</script>
<script src="content-tests.js"></script>
<!--TODO check cache-poisoned resources, transfer-encoding, 3xx redirect
   to resource with matching hash, and cross-origin leakage test as in sec5.3.
   -->
